services/authentication/tokenRequiredDecorator.py
from functools import wraps

from flask import request, jsonify
import jwt
from constants.applicationConstant import SECRET_KEY


def tokenRequired(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        # jwt is passed in the request header
        if 'x-access-token' in request.headers:
            token = request.headers['x-access-token']
        # return 401 if token is not passed
        if not token:
            return jsonify({'message': 'Token is missing !!'}), 401

        try:
            # decoding the payload to fetch the stored details
            data = jwt.decode(token, SECRET_KEY, algorithms="HS256")
            print(data['user'])
        except:
            return jsonify({
                'message': 'Token is invalid !!'
            }), 401
        # returns the current logged in users context to the routes
        return f(data['user'], *args, **kwargs)

    return decorated
