requirements.txt
AutoDotEnv==0.1.3
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.3
paramiko==3.0.0
Pillow==9.5.0
pycparser==2.21
pymongo==4.3.3
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
XlsxWriter==3.0.8

opencv-python~=4.7.0.72
PyJWT~=2.7.0